[PM-42110] refactor: Fold toCopyViewState into SendViewExtensions
Extension files in this codebase are named after the receiver type
(`CipherViewExtensions`, `AttachmentViewExtensions`, `AddEditSendStateExtensions`),
and nothing else splits a second file out per operation. `toCopyViewState` is a
`SendView` extension, so it belongs beside `toViewState` and `toAddEditViewState`
rather than in its own `SendViewCopyExtensions` file.

Move the function and its tests into `SendViewExtensions` and
`SendViewExtensionsTest`, then delete the copy-specific files. The merged test
fixtures drop the duplicate text-type constant, keeping `DEFAULT_TEXT_TYPE`, and
rename `EXPECTED_COMMON` to `COPY_COMMON` so it reads distinctly next to
`DEFAULT_COMMON` now that both live in one file.
